+++ relay-redis/src/scripts/is_rate_limited.lua
@@ -37,19 +37,21 @@ local all_values = redis.call('MGET', unpack(KEYS))
 local results = {}
 local failed = false
 local num_quotas = #KEYS / 2
-for i=0, num_quotas - 1 do
+for i = 0, num_quotas - 1 do
     local k = i * 2 + 1
     local v = i * 4 + 1
 
     local limit = tonumber(ARGV[v])
-    local quantity = tonumber(ARGV[v+2])
-    local over_accept_once = ARGV[v+3]
-    local rejected = false
+    local quantity = tonumber(ARGV[v + 2])
+    local over_accept_once = ARGV[v + 3]
+
+    local main_value = all_values[k] or 0
+    local refund_value = all_values[k + 1] or 0
+    local consumed = main_value - refund_value
+
+    local rejected = false;
     -- limit=-1 means "no limit"
     if limit >= 0 then
-        local main_value = all_values[k] or 0
-        local refund_value = all_values[k + 1] or 0
-        local consumed = main_value - refund_value
         -- Without over_accept_once, we never increment past the limit. if quantity is 0, check instead if we reached limit.
         -- With over_accept_once, we only reject if the previous update already reached the limit.
         -- This way, we ensure that we increment to or past the limit at some point,
@@ -63,14 +65,14 @@ for i=0, num_quotas - 1 do
         end
     end
 
-    if rejected then
-        failed = true
-    end
-    results[i + 1] = rejected
+    failed = failed or rejected
+
+    table.insert(results, rejected)
+    table.insert(results, consumed)
 end
 
 if not failed then
-    for i=0, num_quotas - 1 do
+    for i = 0, num_quotas - 1 do
         local k = i * 2 + 1
         local v = i * 4 + 1