+++ relay-server/src/processing/spans/process.rs
@@ -175,6 +175,7 @@ mod tests {
     use chrono::DateTime;
     use relay_pii::{DataScrubbingConfig, PiiConfig};
     use relay_protocol::SerializableAnnotated;
+    use relay_sampling::evaluation::ReservoirCounters;
 
     use crate::services::projects::project::ProjectInfo;
 
@@ -195,6 +196,7 @@ mod tests {
             ..Default::default()
         }));
         let rate_limits = Box::leak(Box::new(relay_quotas::RateLimits::default()));
+        let reservoir_counters = Box::leak(Box::new(ReservoirCounters::default()));
 
         Context {
             config,
@@ -202,6 +204,7 @@ mod tests {
             project_info,
             rate_limits,
             sampling_project_info: None,
+            reservoir_counters,
         }
     }