+++ relay-server/src/services/upload.rs
@@ -2,106 +2,51 @@
 use std::sync::Arc;
 use std::time::Duration;
 
-use bytes::Bytes;
 use futures::future::BoxFuture;
 use futures::stream::FuturesUnordered;
 use futures::{FutureExt, StreamExt};
+use objectstore_client::{Client, ExpirationPolicy, Session, Usecase};
 use relay_config::UploadServiceConfig;
-use relay_quotas::DataCategory;
-use relay_system::{FromMessage, Interface, NoResponse, Receiver, Recipient, Service};
-use smallvec::smallvec;
-use tokio::sync::Notify;
-use uuid::Uuid;
+use relay_system::{Addr, FromMessage, Interface, NoResponse, Receiver, Service};
 
-use crate::managed::{Counted, Managed, OutcomeError, Quantities, Rejected};
+use crate::constants::DEFAULT_ATTACHMENT_RETENTION;
+use crate::envelope::{Item, ItemType};
+use crate::managed::{OutcomeError, TypedEnvelope};
 use crate::services::outcome::DiscardReason;
+use crate::services::processor::Processed;
+use crate::services::store::{Store, StoreEnvelope};
 use crate::statsd::{RelayCounters, RelayGauges};
 
 use super::outcome::Outcome;
 
-/// Message that requests an attachment upload.
-pub struct UploadAttachment {
-    /// The attachment to be uploaded.
-    pub attachment: Managed<Attachment>,
-
-    /// The return address in case of a successful upload.
-    pub respond_to: Recipient<Managed<UploadedAttachment>, NoResponse>,
+/// Message that requests all attachments of the [`Envelope`](crate::envelope::Envelope) to be uploaded.
+pub struct UploadAttachments {
+    /// The envelope containing attachments for upload.
+    pub envelope: TypedEnvelope<Processed>,
 }
 
-pub enum Upload {
-    Attachment(UploadAttachment),
-}
+impl Interface for UploadAttachments {}
 
-impl Interface for Upload {}
-
-impl FromMessage<UploadAttachment> for Upload {
+impl FromMessage<UploadAttachments> for UploadAttachments {
     type Response = NoResponse;
 
-    fn from_message(message: UploadAttachment, sender: ()) -> Self {
-        Self::Attachment(message)
+    fn from_message(message: UploadAttachments, _sender: ()) -> Self {
+        message
     }
-}
-
-/// The attachment to upload.
-#[derive(Clone, Debug)]
-pub struct Attachment {
-    /// Attachment metadata.
-    pub meta: AttachmentMeta,
-    /// The attachment body.
-    payload: Bytes,
-}
-
-impl Counted for Attachment {
-    fn quantities(&self) -> Quantities {
-        smallvec![
-            (DataCategory::Attachment, self.payload.len()),
-            (DataCategory::AttachmentItem, 1),
-        ]
-    }
-}
-
-/// The result of a successful attachment upload.
-///
-/// This is tracked so that the recipient of the success message can emit outcomes for the
-/// attachment.
-#[derive(Clone, Debug)]
-pub struct UploadedAttachment {
-    meta: AttachmentMeta,
-    uploaded_bytes: usize,
-}
-
-impl Counted for UploadedAttachment {
-    fn quantities(&self) -> Quantities {
-        smallvec![
-            (DataCategory::Attachment, self.uploaded_bytes),
-            (DataCategory::AttachmentItem, 1),
-        ]
-    }
-}
-
-/// Metadata of the attachment (stub).
-#[derive(Clone, Debug)]
-pub struct AttachmentMeta {
-    attachment_id: Option<String>,
-    // TODO: more fields
 }
 
 /// Errors that can occur when trying to upload an attachment.
 #[derive(Debug)]
 pub enum Error {
-    LoadShed,
     Timeout,
     UploadFailed,
-    KeyMismatch,
 }
 
 impl Error {
     fn as_str(&self) -> &'static str {
         match self {
-            Error::LoadShed => "load_shed",
             Error::Timeout => "timeout",
             Error::UploadFailed => "upload_failed",
-            Error::KeyMismatch => "key_mismatch",
         }
     }
 }
@@ -118,56 +63,75 @@ impl OutcomeError for Error {
 ///
 /// Accepts upload requests and maintains a list of concurrent uploads.
 pub struct UploadService {
-    pending_requests: FuturesUnordered<BoxFuture<'static, Result<(), Rejected<Error>>>>,
+    pending_requests: FuturesUnordered<BoxFuture<'static, ()>>,
     max_concurrent_requests: usize,
-    timeout: Duration,
+    inner: Arc<UploadServiceInner>,
 }
 
 impl UploadService {
-    pub fn new(config: &UploadServiceConfig) -> Self {
+    pub fn new(
+        config: &UploadServiceConfig,
+        store: Option<Addr<Store>>,
+    ) -> anyhow::Result<Option<Self>> {
+        let Some(store) = store else { return Ok(None) };
         let UploadServiceConfig {
+            objectstore_url,
             max_concurrent_requests,
             timeout,
         } = config;
-        Self {
+        let Some(objectstore_url) = objectstore_url else {
+            return Ok(None);
+        };
+
+        let objectstore_client = Client::builder(objectstore_url).build()?;
+        let attachments_usecase = Usecase::new("attachments")
+            .with_expiration_policy(ExpirationPolicy::TimeToLive(DEFAULT_ATTACHMENT_RETENTION));
+
+        let inner = UploadServiceInner {
+            timeout: Duration::from_secs(*timeout),
+
+            store,
+
+            objectstore_client,
+            attachments_usecase,
+        };
+
+        Ok(Some(Self {
             pending_requests: FuturesUnordered::new(),
             max_concurrent_requests: *max_concurrent_requests,
-            timeout: Duration::from_secs(*timeout),
-        }
+            inner: Arc::new(inner),
+        }))
     }
 
-    fn handle_message(&mut self, message: Upload) {
-        let Upload::Attachment(UploadAttachment {
-            attachment,
-            respond_to,
-        }) = message;
+    fn handle_message(&self, message: UploadAttachments) {
+        let UploadAttachments { envelope } = message;
         if self.pending_requests.len() >= self.max_concurrent_requests {
             // Load shed to prevent backlogging in the service queue and affecting other parts of Relay.
             // We might want to have a less aggressive mechanism in the future.
-            count_upload(Err(attachment.reject_err(Error::LoadShed)));
+
+            // for now, this will just forward to the store endpoint without uploading attachments.
+            self.inner.store.send(StoreEnvelope { envelope });
             return;
         }
 
-        self.pending_requests
-            .push(handle_upload(self.timeout, attachment, respond_to).boxed());
+        let inner = self.inner.clone();
+        let future = async move { inner.handle_envelope(envelope).await };
+        self.pending_requests.push(future.boxed());
     }
 }
 
 impl Service for UploadService {
-    type Interface = Upload;
+    type Interface = UploadAttachments;
 
     async fn run(mut self, mut rx: Receiver<Self::Interface>) {
         relay_log::info!("Upload service started");
         loop {
             relay_log::trace!("Upload loop iteration");
             tokio::select! {
-                //Bias towards handling responses so that there's space for new incoming requests.
+                // Bias towards handling responses so that there's space for new incoming requests.
                 biased;
 
-                Some(result) = self.pending_requests.next() => {
-                    relay_log::trace!("One upload has finished");
-                    count_upload(result);
-                }
+                Some(_) = self.pending_requests.next() => {},
                 Some(message) = rx.recv() => self.handle_message(message),
 
                 else => break,
@@ -181,92 +145,78 @@ impl Service for UploadService {
     }
 }
 
-fn count_upload(result: Result<(), Rejected<Error>>) {
-    let result_msg = match result {
-        Ok(()) => "success",
-        Err(e) => e.into_inner().as_str(),
-    };
-    relay_statsd::metric!(
-        counter(RelayCounters::AttachmentUpload) += 1,
-        result = result_msg
-    );
+struct UploadServiceInner {
+    timeout: Duration,
+    store: Addr<Store>,
+
+    objectstore_client: Client,
+    attachments_usecase: Usecase,
 }
 
-/// Spend a limited time trying to upload an attachment, and emit outcomes if this fails.
-///
-/// Returns an [`UploadedAttachment`] if the upload was successful.
-async fn handle_upload(
-    timeout: Duration,
-    attachment: Managed<Attachment>,
-    respond_to: Recipient<Managed<UploadedAttachment>, NoResponse>,
-) -> Result<(), Rejected<Error>> {
-    relay_log::trace!("Starting upload");
-    let key = attachment.meta.attachment_id.as_deref();
-    let new_key = tokio::time::timeout(timeout, async { upload(key, &attachment.payload).await })
-        .await
-        .map_err(|_elapsed| attachment.reject_err(Error::Timeout))?
-        .map_err(|_error| attachment.reject_err(Error::UploadFailed))?;
+impl UploadServiceInner {
+    /// Uploads all attachments belonging to the given envelope.
+    ///
+    /// This mutates the attachment items in-place, setting their `stored_key` field to the key
+    /// in objectstore.
+    async fn handle_envelope(&self, mut envelope: TypedEnvelope<Processed>) -> () {
+        let scoping = envelope.scoping();
+        let session = self
+            .attachments_usecase
+            .for_project(scoping.organization_id.value(), scoping.project_id.value())
+            .session(&self.objectstore_client);
 
-    if key.is_some_and(|key| key != new_key) {
-        return Err(attachment.reject_err(Error::KeyMismatch));
+        let attachments = envelope
+            .envelope_mut()
+            .items_mut()
+            .filter(|item| *item.ty() == ItemType::Attachment);
+        let upload_results = if let Ok(session) = session {
+            futures::future::join_all(
+                attachments.map(|attachment| self.handle_attachment(&session, attachment)),
+            )
+            .await
+        } else {
+            // if we have any kind of error constructing the upload session, we mark all attachments as failed
+            attachments.map(|_| Err(Error::UploadFailed)).collect()
+        };
+
+        for result in upload_results {
+            let result_msg = match result {
+                Ok(()) => "success",
+                Err(error) => error.as_str(),
+            };
+            relay_statsd::metric!(
+                counter(RelayCounters::AttachmentUpload) += 1,
+                result = result_msg
+            );
+        }
+
+        // last but not least, forward the envelope to the store endpoint
+        self.store.send(StoreEnvelope { envelope });
     }
 
-    let uploaded_attachment = attachment.map(|Attachment { mut meta, payload }, _| {
-        meta.attachment_id = Some(new_key);
-        UploadedAttachment {
-            meta,
-            uploaded_bytes: payload.len(),
+    async fn handle_attachment(
+        &self,
+        session: &Session,
+        attachment: &mut Item,
+    ) -> Result<(), Error> {
+        // we are not storing zero-size attachments in objectstore
+        if attachment.is_empty() {
+            return Ok(());
         }
-    });
+        relay_log::trace!("Starting attachment upload");
+        let future = async {
+            let result = session.put(attachment.payload()).send().await?;
+            Ok(result.key)
+        };
 
-    respond_to.send(uploaded_attachment);
-    relay_log::trace!("Finished upload");
-    Ok(())
-}
+        let stored_key = tokio::time::timeout(self.timeout, future)
+            .await
+            .map_err(|_elapsed| Error::Timeout)?
+            .map_err(|_error: objectstore_client::Error| Error::UploadFailed)?;
 
-/// Uploads the payload to the objectstore and returns the key under which it is stored.
-///
-/// - If `key` is `None`, the objectstore will assign a key.
-/// - If `key` is not `None`, write to objectstore with the given key.
-async fn upload(key: Option<&str>, payload: &[u8]) -> Result<String, ()> {
-    // TODO: call objectstore
-    Ok(key.unwrap_or_default().to_owned())
-}
+        attachment.set_stored_key(stored_key);
+        relay_log::trace!("Finished attachment upload");
 
-#[cfg(test)]
-mod tests {
-    use relay_redis::redis::FromRedisValue;
-    use relay_system::Addr;
-
-    use super::*;
-
-    #[tokio::test]
-    async fn test_basic() {
-        let upload_service = UploadService::new(&UploadServiceConfig {
-            max_concurrent_requests: 2,
-            timeout: 1,
-        })
-        .start_detached();
-
-        let (tx, mut rx) = tokio::sync::mpsc::unbounded_channel();
-
-        let (attachment, mut managed_handle) = Managed::for_test(Attachment {
-            meta: AttachmentMeta {
-                attachment_id: Some("my_key".to_owned()),
-            },
-            payload: Bytes::from("hello world"),
-        })
-        .build();
-
-        upload_service.send(UploadAttachment {
-            attachment,
-            respond_to: Recipient::<_, NoResponse>::new(tx),
-        });
-        let uploaded = rx.recv().await.unwrap();
-        assert_eq!(uploaded.meta.attachment_id.as_deref(), Some("my_key"));
-        uploaded.accept(|_| {});
-
-        drop(upload_service);
-        assert!(rx.recv().await.is_none());
+        Ok(())
     }
 }